refactor(): tweaks to deal with integration of older commits
