+++ src/core/geometry/shape2/validate.js
@@ -36,7 +36,7 @@ const isCAGValid = function (CAG) {
   }
 }
 
-const isSelfIntersecting = function (cag, debug) {
+/* const isSelfIntersecting = function (cag, debug) {
   let numsides = cag.sides.length
   for (let i = 0; i < numsides; i++) {
     let side0 = cag.sides[i]
@@ -49,51 +49,6 @@ const isSelfIntersecting = function (cag, debug) {
     }
   }
   return false
-}
+} */
 
-/** Check if the point stay inside the CAG shape
-* ray-casting algorithm based on :
-* https://github.com/substack/point-in-polygon/blob/master/index.js
-* http://www.ecse.rp1.edu/Homepages/wrf/Research/Short_Notes/pnpoly.html
-* originaly writed for https://github.com/lautr3k/SLAcer.js/blob/dev/js/slacer/slicer.js#L82
-* @param {CAG} cag - CAG object
-* @param {Object} p0 - Vertex2 like object
-* @returns {Boolean}
-*/
-const hasPointInside = function (cag, p0) {
-  let p1 = null
-  let p2 = null
-  let inside = false
-  cag.sides.forEach(side => {
-    p1 = side.vertex0.pos
-    p2 = side.vertex1.pos
-    if (hasPointInside.c1(p0, p1, p2) && hasPointInside.c2(p0, p1, p2)) {
-      inside = !inside
-    }
-  })
-  return inside
-}
-
-hasPointInside.c1 = (p0, p1, p2) => (p1.y > p0.y) !== (p2.y > p0.y)
-hasPointInside.c2 = (p0, p1, p2) => (p0.x < (p2.x - p1.x) * (p0.y - p1.y) / (p2.y - p1.y) + p1.x)
-
-/** Check if all points from one CAG stay inside another CAG
-* @param {CAG} cag1 - CAG object
-* @param {Object} cag2 - CAG object
-* @returns {Boolean}
-*/
-const contains = function (cag1, cag2) {
-  for (let i = 0, il = cag2.sides.length; i < il; i++) {
-    if (!hasPointInside(cag1, cag2.sides[i].vertex0.pos)) {
-      return false
-    }
-  }
-  return true
-}
-
-module.exports = {
-  isCAGValid,
-  isSelfIntersecting,
-  hasPointInside,
-  contains
-}
+module.exports = isCAGValid
